refactor: RepositoryBaseTest로 이름 변경
